Updates to reorder-files.rb and some other notes

+++ bin/reorder-files.rb
@@ -1,44 +1,45 @@
-require 'fileutils'
-require 'taglib'
+# frozen_string_literal: true
 
-files = Dir['split/*']
-new_dir = 'renamed'
+require "fileutils"
+require "taglib"
+
+if ARGV.length < 2
+  raise "Too few arguments"
+  exit
+end
+
+src_dir = ARGV[0]
+track_postfix = ARGV[1]
+dry_run = ARGV[2]
+raise "Dir.exist?(#{src_dir}) fail" unless Dir.exist?(src_dir)
+files = Dir["#{ARGV.first}/*"]
+new_dir = "#{ARGV.first}-renamed"
 Dir.mkdir(new_dir) unless Dir.exist?(new_dir)
 
 files.sort.each_with_index do |f, i|
   track = i + 1
-  new_title = "#{sprintf('%03d', track)} - Right Stuff"
+  new_title = "#{format('%03d', track)} - #{track_postfix}"
   new_file_path = "#{new_dir}/#{new_title}.mp3"
 
   puts f
   puts new_file_path
-  FileUtils.cp(f, new_file_path)
-  TagLib::MPEG::File.open(new_file_path) do |file|
-    tag2 = file.id3v2_tag
-    tag1 = file.id3v1_tag
+  if dry_run
+    puts "--- DRY RUN -- nothing copied"
+  else
+    FileUtils.cp(f, new_file_path)
+    TagLib::MPEG::File.open(new_file_path) do |file|
+      tag2 = file.id3v2_tag
+      tag1 = file.id3v1_tag
 
-    puts tag1.title
-    puts tag2.track
-    puts tag1.title
-    puts tag2.track
-    tag1.title = new_title
-    tag1.track = track
-    tag2.title = new_title
-    tag2.track = track
-    file.save
+      puts tag1.title
+      puts tag2.track
+      puts tag1.title
+      puts tag2.track
+      tag1.title = new_title
+      tag1.track = track
+      tag2.title = new_title
+      tag2.track = track
+      file.save
+    end
   end
 end
-
-# Load a tag from a file
-# tag = ID3Lib::Tag.new('talk.mp3')
-# files = Dir['renamed/*']
-#
-# file_groups = files.sort.each_slice(11).to_a
-# file_groups.each_with_index do |file_group, i|
-#    new_dir = "Part#{sprintf '%02d', (i + 1)}"
-#    Dir.mkdir(new_dir) unless Dir.exist?(new_dir)
-#    file_group.each do|f|
-#      FileUtils.cp(f, new_dir)
-#    end
-#   puts new_dir
-# end
